blood_donation_api/requirements.txt
# Requires Python 3.11 or 3.12 (not 3.14 — pydantic-core has no wheels for 3.14).
fastapi==0.111.0
uvicorn[standard]==0.29.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.7.0
httpx==0.27.0
slowapi==0.1.9
haversine==2.8.1
